ARM ReOrg Compiles with no error
