+++ packages/typespec-azure-resource-manager/lib/common-types/common-types.tsp
@@ -1,10 +1,18 @@
 import "@typespec/versioning";
+import "./common-types.tsp";
+import "./types-ref.tsp";
+import "./managed-identity-ref.tsp";
+import "./private-links-ref.tsp";
+import "./customer-managed-keys-ref.tsp";
+import "./extended-location-ref.tsp";
+import "./commontypes.private.decorators.tsp";
 
 using TypeSpec.Versioning;
 
+@versioned(Versions)
 namespace Azure.ResourceManager.CommonTypes;
 
-@Private.armCommonTypesVersions
+@CommonTypes.Private.armCommonTypesVersions
 @doc("The Azure Resource Manager common-types versions.")
 enum Versions {
   @doc("The Azure Resource Manager v3 common types.")
