+++ packages/typespec-azure-resource-manager/lib/common-types/commontypes.private.decorators.tsp
@@ -1,0 +1,43 @@
+import "../../dist/src/commontypes.private.decorators.js";
+
+namespace Azure.ResourceManager.CommonTypes.Private;
+
+using TypeSpec.Reflection;
+
+/**
+ * Describes the shape of model literals accepted by the `version` parameter of
+ * the `armCommonDefinition` and `armCommonParameter` decorators.
+ */
+alias ArmCommonTypeVersionSpec = {
+  version: string | EnumMember;
+  isDefault: boolean;
+};
+
+/**
+ * @param definitionName Definition name
+ * @param version Azure Resource Manager Version
+ * @param referenceFile Reference file
+ */
+extern dec armCommonDefinition(
+  target: Model,
+  definitionName?: valueof string,
+  version?: valueof EnumMember | ArmCommonTypeVersionSpec | string,
+  referenceFile?: valueof string
+);
+
+/**
+ * @param definitionName Definition name
+ * @param version Azure Resource Manager Version
+ * @param referenceFile Reference file
+ */
+extern dec armCommonParameter(
+  target: ModelProperty,
+  definitionName?: valueof string,
+  version?: valueof EnumMember | ArmCommonTypeVersionSpec | string,
+  referenceFile?: valueof string
+);
+
+/**
+ * Marks an enum as representing the valid `common-types` versions.
+ */
+extern dec armCommonTypesVersions(target: Enum);
