+++ packages/typespec-azure-resource-manager/lib/common-types/customer-managed-keys-ref.tsp
@@ -1,0 +1,14 @@
+import "./customer-managed-keys.tsp";
+
+namespace Azure.ResourceManager.CommonTypes.Private;
+
+@@armCommonDefinition(EncryptionConfiguration,
+  "encryption",
+  Azure.ResourceManager.CommonTypes.Versions.v4,
+  "customermanagedkeys.json"
+);
+@@armCommonDefinition(EncryptionConfiguration,
+  "encryption",
+  Azure.ResourceManager.CommonTypes.Versions.v5,
+  "customermanagedkeys.json"
+);
