+++ packages/typespec-azure-resource-manager/lib/common-types/extended-location-ref.tsp
@@ -1,0 +1,1 @@
+import "./extended-location.tsp";
