+++ packages/typespec-azure-resource-manager/lib/common-types/private-links.tsp
@@ -1,27 +1,10 @@
-import "@typespec/openapi";
-import "@typespec/http";
-import "@typespec/rest";
-import "@typespec/versioning";
-import "@azure-tools/typespec-azure-core";
-
 using TypeSpec.Http;
-using TypeSpec.OpenAPI;
 using TypeSpec.Versioning;
-using Azure.ResourceManager.Private;
 
-namespace Azure.ResourceManager;
+namespace Azure.ResourceManager.CommonTypes;
 
 /** The private endpoint resource */
-@armCommonDefinition(
-  "PrivateEndpoint",
-  #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true },
-  "privatelinks.json"
-)
-@armCommonDefinition(
-  "PrivateEndpoint",
-  Azure.ResourceManager.CommonTypes.Versions.v5,
-  "privatelinks.json"
-)
+@added(CommonTypes.Versions.v4)
 model PrivateEndpoint {
   /** The resource identifier for private endpoint */
   id?: Azure.Core.armResourceIdentifier<[
@@ -32,22 +15,14 @@ model PrivateEndpoint {
 }
 
 /** The private endpoint connection resource */
+@added(CommonTypes.Versions.v4)
 model PrivateEndpointConnection extends Azure.ResourceManager.Foundations.ProxyResource {
   /** The private endpoint connection properties */
   properties?: PrivateEndpointConnectionProperties;
 }
 
 /** Properties of he private endpoint connection resource */
-@armCommonDefinition(
-  "PrivateEndpointConnectionProperties",
-  #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true },
-  "privatelinks.json"
-)
-@armCommonDefinition(
-  "PrivateEndpointConnectionProperties",
-  Azure.ResourceManager.CommonTypes.Versions.v5,
-  "privatelinks.json"
-)
+@added(CommonTypes.Versions.v4)
 model PrivateEndpointConnectionProperties {
   /** The group identifiers for the private endpoint resource */
   @visibility("read")
@@ -64,16 +39,7 @@ model PrivateEndpointConnectionProperties {
 }
 
 /** A collection of information about the state of the connection between service consumer and provider. */
-@armCommonDefinition(
-  "PrivateLinkServiceConnectionState",
-  #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true },
-  "privatelinks.json"
-)
-@armCommonDefinition(
-  "PrivateLinkServiceConnectionState",
-  Azure.ResourceManager.CommonTypes.Versions.v5,
-  "privatelinks.json"
-)
+@added(CommonTypes.Versions.v4)
 model PrivateLinkServiceConnectionState {
   /** Indicates whether the connection has been Approved/Rejected/Removed by the owner of the service. */
   status?: PrivateEndpointServiceConnectionStatus;
@@ -86,6 +52,7 @@ model PrivateLinkServiceConnectionState {
 }
 
 /** The provisioning state of the connection */
+@added(CommonTypes.Versions.v4)
 union PrivateEndpointConnectionProvisioningState {
   ResourceProvisioningState,
 
@@ -97,6 +64,7 @@ union PrivateEndpointConnectionProvisioningState {
 }
 
 /** The private endpoint connection status */
+@added(CommonTypes.Versions.v4)
 union PrivateEndpointServiceConnectionStatus {
   /** Connectionaiting for approval or rejection */
   Pending: "Pending",
@@ -110,65 +78,14 @@ union PrivateEndpointServiceConnectionStatus {
   string,
 }
 
-model PrivateLinkResource extends Azure.ResourceManager.Foundations.ProxyResource {
-  /** Properties of the private link resource. */
-  properties?: PrivateLinkResourceProperties;
-}
-
-/** Properties of a private link resource. */
-#suppress "@azure-tools/typespec-azure-resource-manager/arm-resource-provisioning-state" "Matches current common code"
-@armCommonDefinition(
-  "PrivateLinkResourceProperties",
-  #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true },
-  "privatelinks.json"
-)
-@armCommonDefinition(
-  "PrivateLinkResourceProperties",
-  Azure.ResourceManager.CommonTypes.Versions.v5,
-  "privatelinks.json"
-)
-model PrivateLinkResourceProperties {
-  /** The private link resource group id. */
-  @visibility("read")
-  groupId?: string;
-
-  /** The private link resource required member names. */
-  @visibility("read")
-  requiredMembers?: string[];
-
-  /** The private link resource private link DNS zone name. */
-  requiredZoneNames?: string[];
-}
-
-/** List of private endpoint connections associated with the specified resource. */
-model PrivateEndpointConnectionResourceListResult {
-  /** Array of private endpoint connections */
-  value?: PrivateEndpointConnection[];
-}
-
-/** A list of private link resources. */
-model PrivateLinkResourceListResult {
-  /** Array of private link resources */
-  value?: PrivateLinkResource[];
-}
-
 /**
  * The name of the private endpoint connection associated with the Azure resource.
  * @template Segment The resource type name for private endpoint connections (default is privateEndpointConnections)
  */
+@added(CommonTypes.Versions.v4)
 model PrivateEndpointConnectionParameter<Segment extends valueof string = "privateEndpointConnections"> {
   /** The name of the private endpoint connection associated with the Azure resource. */
   @path
-  @armCommonParameter(
-    "PrivateEndpointConnectionName",
-    #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true },
-    "privatelinks.json"
-  )
-  @armCommonParameter(
-    "PrivateEndpointConnectionName",
-    Azure.ResourceManager.CommonTypes.Versions.v5,
-    "privatelinks.json"
-  )
   @TypeSpec.Rest.segment(Segment)
   @key("privateEndpointConnectionName")
   name: string;
@@ -178,6 +95,7 @@ model PrivateEndpointConnectionParameter<Segment extends valueof string = "priva
  * The name of the private link associated with the Azure resource.
  * @template Segment The resource type name for private links (default is privateLinkResources)
  */
+@added(CommonTypes.Versions.v4)
 model PrivateLinkResourceParameter<Segment extends valueof string = "privateLinkResources"> {
   /** The name of the private link associated with the Azure resource. */
   @path
