+++ packages/typespec-azure-resource-manager/lib/private.decorators.tsp
@@ -15,39 +15,6 @@ extern dec armUpdateProviderNamespace(target: Reflection.Operation);
 extern dec assignProviderNameValue(target: ModelProperty, resource: Model);
 
 /**
- * Describes the shape of model literals accepted by the `version` parameter of
- * the `armCommonDefinition` and `armCommonParameter` decorators.
- */
-alias ArmCommonTypeVersionSpec = {
-  version: string | EnumMember;
-  isDefault: boolean;
-};
-
-/**
- * @param definitionName Definition name
- * @param version Azure Resource Manager Version
- * @param referenceFile Reference file
- */
-extern dec armCommonDefinition(
-  target: Model,
-  definitionName?: valueof string,
-  version?: valueof EnumMember | ArmCommonTypeVersionSpec | string,
-  referenceFile?: valueof string
-);
-
-/**
- * @param definitionName Definition name
- * @param version Azure Resource Manager Version
- * @param referenceFile Reference file
- */
-extern dec armCommonParameter(
-  target: ModelProperty,
-  definitionName?: valueof string,
-  version?: valueof EnumMember | ArmCommonTypeVersionSpec | string,
-  referenceFile?: valueof string
-);
-
-/**
  * This decorator is used to identify Azure Resource Manager resource types and extract their
  * metadata.  It is *not* meant to be used directly by a spec author, it instead
  * gets implicitly applied when the spec author defines a model type in this form:
@@ -77,11 +44,6 @@ extern dec resourceBaseParametersOf(target: Model, propertyName: Model);
  * @param values Values
  */
 extern dec resourceParameterBaseFor(target: ModelProperty, values: unknown[]);
-
-/**
- * Marks an enum as representing the valid `common-types` versions.
- */
-extern dec armCommonTypesVersions(target: Enum);
 
 /**
  * Provides default name decoration on resource name property with
