+++ packages/typespec-azure-resource-manager/lib/responses.tsp
@@ -3,6 +3,7 @@ using TypeSpec.Rest;
 using TypeSpec.OpenAPI;
 using Azure.ResourceManager.Foundations;
 using Azure.ResourceManager.Private;
+using Azure.ResourceManager.CommonTypes.Private;
 
 namespace Azure.ResourceManager;
